Remove some warnings

+++ src/tools/ibex_KCoreGraph.cpp
@@ -59,7 +59,7 @@ void KCoreGraph::remove_vertex(const int idvert) {
 	while (!vn->empty()) {
 		val = vn->head();
 		neighbourhoods.at(val)->remove(idvert);
-		if (neighbourhoods.at(val)->size < k) {
+		if (((int)(neighbourhoods.at(val)->size)) < k) {
 			if (!tbr->contain(val)) tbr->add(val);
 		}
 		vn->remove(val);
@@ -92,7 +92,7 @@ void KCoreGraph::apply_coreness() {
 	
 	/* Not optimal, obviously */
 	for (unsigned int i=0; i<neighbourhoods.size(); i++) {
-		if ((neighbourhoods.at(i) != NULL) && (neighbourhoods.at(i)->size < k)) {
+		if ((neighbourhoods.at(i) != NULL) && (((int)(neighbourhoods.at(i)->size)) < k)) {
 			remove_vertex(i);
 			allid->remove(i);
 			propagate();
